Strategist email: per-client cadence + recipients; Brain polish; bug/security sweep fixes
Per-client Strategist Intelligence email control (Gary):
- intel_briefs gains email_schedule (off/daily/weekly, default weekly) +
  email_recipients (jsonb); applied to Neon
- /api/studio/intel/schedule GET/POST reads + writes the per-brain setting
- IntelEmailControl on the Strategist page: 3-state cadence + add/remove
  recipient chips, saved on change
- daily-intel cron honours each brain's cadence (off = no run/no spend, weekly
  = Monday only, daily = weekdays; manual ?clientId= overrides) and mails the
  brain's own recipients (empty falls back to the platform default)
- email copy is cadence-driven: heading, strapline and footer now agree
  (no more "Weekly" strapline over a "Daily" heading); emails are the
  mobile-first shell already
- footer signs by role carrying the client name (AI Research Strategist MTN
  MoMo), not "Sami"; the platform "Review and accept" CTA is removed since the
  digest can land in a client inbox with no login

Brain page polish:
- Website panel shows the actual last-crawl DATE and a subtle 30-day
  re-crawl nudge (per row + a one-line summary when any are due)
- consistent section iconography across Knowledge sources, coverage,
  What the brain knows, and the Brand library

Bug + security sweep (3 expert passes) fixes:
- email esc() now escapes quotes too: source URLs land in href="..." and a
  bare quote was an attribute-injection gap (insert-time http(s) filter stays)
- chunk_count counts cast ::int at source (getBrain/listBrains/listSources) +
  defensive Number() coercions, killing the "01372" string-concat class and
  restoring the empty-brain onboarding hero
- copy: strategist empty-queue line no longer says "Friday"; removed an en dash
  and an unused import

Core client_id brain isolation and auth gating verified clean (no
CRITICAL/HIGH). Remaining LOW/iteration-2 items reported separately.

+++ lib/brains.ts
@@ -20,8 +20,8 @@ function slugify(name: string): string {
 export async function listBrains(): Promise<Brain[]> {
   return (await db().query(
     `select c.id, c.name, c.slug, c.status, c.brand, c.created_at,
-            (select count(*) from knowledge_chunks k where k.client_id = c.id)  as chunk_count,
-            (select count(*) from knowledge_sources s where s.client_id = c.id) as source_count
+            (select count(*)::int from knowledge_chunks k where k.client_id = c.id)  as chunk_count,
+            (select count(*)::int from knowledge_sources s where s.client_id = c.id) as source_count
      from clients c order by c.created_at desc`,
   )) as Brain[];
 }
@@ -29,7 +29,7 @@ export async function listBrains(): Promise<Brain[]> {
 export async function getBrain(id: string): Promise<Brain | null> {
   const rows = (await db().query(
     `select id, name, slug, status, brand, created_at,
-            (select count(*) from knowledge_chunks k where k.client_id = clients.id) as chunk_count
+            (select count(*)::int from knowledge_chunks k where k.client_id = clients.id) as chunk_count
      from clients where id = $1`,
     [id],
   )) as Brain[];
@@ -69,7 +69,7 @@ export type KnowledgeSource = {
 export async function listSources(clientId: string): Promise<KnowledgeSource[]> {
   return (await db().query(
     `select s.id, s.client_id, s.type, s.uri, s.status, s.error, s.last_synced_at,
-            (select count(*) from knowledge_chunks k where k.source_id = s.id) as chunk_count
+            (select count(*)::int from knowledge_chunks k where k.source_id = s.id) as chunk_count
      from knowledge_sources s where s.client_id = $1 order by s.id desc`,
     [clientId],
   )) as KnowledgeSource[];
